Skip adding donation records for missing users
addDonation may encounter historical deposit events for users who have
since left the guild.  If the former guild member was never on the
roster when AMT was installed, then no records will exist for them.

Skip processing their donations in this case.  This is similar to the
handling in updateJoinDate, except we assume the guild itself exists.